playground/python/jenkins/requirements.txt
GitPython
requests
markdown